Reorganise map files (#45)
* Rename fluxy.operators.flux.py to fluxy.operators.flux_map_resample

* Rename fluxy.operators.flux_resample.py to fluxy.operators.flux_timeseries_resample

* Move flux_map_resample.define_var_plot.py and select.prepare_flux_map_data to flux_prepare_map_data + renaming prepare_flux_map_data to prepare_data.
