+++ fluxy/plots/flux_map.py
@@ -20,8 +20,8 @@ from fluxy.plots.utils import (
     compute_boundary_geometry,
     plot_country_borders,
 )
-from fluxy.operators.flux import get_flux_mean, average_over_period, define_var_plot
-from fluxy.operators.select import prepare_flux_map_data
+from fluxy.operators.flux_prepare_map_data import define_var_plot, prepare_data
+from fluxy.operators.flux_map_resample import get_flux_mean, average_over_period
 
 
 def plot_flux_map(
@@ -312,7 +312,7 @@ def plot_flux_map_model_comparison(
 
     # Prepare datasets
     ds_dict = {k: v for k, v in ds_all.items() if k in models}
-    ds_dict = prepare_flux_map_data(ds_dict)
+    ds_dict = prepare_data(ds_dict)
     ds_dict["diff"] = ds_dict[models[1]] - ds_dict[models[0]]
     ds_dict["diff"].attrs["frequency"] = ds_dict[models[0]].attrs["frequency"]  # Copy attributes from models[0]
     for v in ds_dict["diff"].data_vars:
@@ -497,7 +497,7 @@ def plot_flux_map_over_time(
         )
 
     # Prepare datasets and average over given periods
-    ds_dict = prepare_flux_map_data(ds_all, plot_combined)
+    ds_dict = prepare_data(ds_all, plot_combined)
     ds_chopby = {}
     for key, ds in ds_dict.items():
         ds_chopby[key], time_labels = average_over_period(ds, dt, chop_by)
